LectureDetailPage, LectureDetailCustomPage 통합
