+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/home/timetable/TimetablePage.kt
@@ -272,9 +272,7 @@ fun TimeTable(
                             for (lecture in lectures) {
                                 if (lecture.contains(day, time)) {
                                     lectureDetailViewModel.initializeEditingLectureDetail(lecture)
-                                    if (lecture.isCustom) {
-                                        navigator.navigate(NavigationDestination.LectureDetailCustom)
-                                    } else navigator.navigate(NavigationDestination.LectureDetail)
+                                    navigator.navigate(NavigationDestination.LectureDetail)
                                     break
                                 }
                             }
