+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/lecture_detail/LectureDetailPage.kt
@@ -23,9 +23,11 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
@@ -36,16 +38,20 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.wafflestudio.snutt2.R
 import com.wafflestudio.snutt2.components.compose.*
 import com.wafflestudio.snutt2.data.TimetableColorTheme
+import com.wafflestudio.snutt2.lib.*
 import com.wafflestudio.snutt2.lib.android.webview.CloseBridge
 import com.wafflestudio.snutt2.lib.android.webview.WebViewContainer
 import com.wafflestudio.snutt2.lib.data.SNUTTStringUtils
+import com.wafflestudio.snutt2.lib.network.dto.core.ClassTimeDto
 import com.wafflestudio.snutt2.lib.network.dto.core.ColorDto
 import com.wafflestudio.snutt2.ui.SNUTTColors
 import com.wafflestudio.snutt2.ui.SNUTTTypography
 import com.wafflestudio.snutt2.ui.isDarkMode
 import com.wafflestudio.snutt2.views.*
 import com.wafflestudio.snutt2.views.logged_in.home.reviews.ReviewWebView
+import com.wafflestudio.snutt2.views.logged_in.home.search.lectureApiWithOverlapDialog
 import com.wafflestudio.snutt2.views.logged_in.home.settings.UserViewModel
+import com.wafflestudio.snutt2.views.logged_in.home.timetable.Defaults
 import kotlinx.coroutines.*
 
 @OptIn(ExperimentalMaterialApi::class)
@@ -57,7 +63,7 @@ fun LectureDetailPage() {
     val navController = LocalNavController.current
     val apiOnProgress = LocalApiOnProgress.current
     val apiOnError = LocalApiOnError.current
-    val pageController = LocalHomePageController.current
+    val focusManager = LocalFocusManager.current
 
     // share viewModel
     val backStackEntry = remember(navController.currentBackStackEntry) {
@@ -69,33 +75,58 @@ fun LectureDetailPage() {
     val editMode by vm.editMode.collectAsState()
     val editingLectureDetail by vm.editingLectureDetail.collectAsState()
     val currentTable by vm.currentTable.collectAsState()
+    val tableColorTheme = currentTable?.theme ?: TimetableColorTheme.SNUTT
+    val isCustom = editingLectureDetail.isCustom
 
     var deleteLectureDialogState by remember { mutableStateOf(false) }
     var resetLectureDialogState by remember { mutableStateOf(false) }
     var editExitDialogState by remember { mutableStateOf(false) }
-    BackHandler(enabled = editMode) {
-        editExitDialogState = true
-    }
-    BackHandler(enabled = viewMode) {
-        vm.setViewMode(false)
-        navController.popBackStack()
-    }
+    var editingClassTimeIndex by remember { mutableStateOf(0) }
+    var deleteTimeDialogState by remember { mutableStateOf(false) }
+    var lectureOverlapDialogState by remember { mutableStateOf(false) }
+    var lectureOverlapDialogMessage by remember { mutableStateOf("") }
 
+    /* 바텀시트 관련 */
     val sheetState = rememberModalBottomSheetState(initialValue = ModalBottomSheetValue.Hidden)
-    var sheetContent by remember {
+    var bottomSheetContent by remember {
         mutableStateOf<@Composable ColumnScope.() -> Unit>({
             Box(modifier = Modifier.size(1.dp))
         })
     }
-    val sheetContentSetter: (@Composable ColumnScope.() -> Unit) -> Unit = {
-        sheetContent = it
+    val bottomSheetContentSetter: (@Composable ColumnScope.() -> Unit) -> Unit = {
+        bottomSheetContent = it
     }
     LaunchedEffect(sheetState.isVisible) {
         if (!sheetState.isVisible) {
-            sheetContent = { Box(modifier = Modifier.size(1.dp)) }
+            bottomSheetContent = { Box(modifier = Modifier.size(1.dp)) }
         }
     }
 
+    /* 뒤로가기 핸들링 */
+    BackHandler {
+        if (sheetState.isVisible) {
+            scope.launch { sheetState.hide() }
+        } else if (editMode) {
+            if (vm.isAddMode()) { // 새 커스텀 강의 추가일 때는 뒤로가기 하면 바로 나가기
+                navController.popBackStack()
+            } else editExitDialogState = true
+        } else if (viewMode) {
+            vm.setViewMode(false)
+            navController.popBackStack()
+        } else {
+            navController.popBackStack()
+        }
+    }
+
+    /* TODO (진행중)
+     * 시간 및 장소 item 추가했을 때 애니메이션 적용하기 (LazyColumn 의 기능 모방)
+     * 추가시 애니메이션은 되는데 삭제시는 방법을 고민중
+     */
+    var classTimeAnimationState by remember {
+        mutableStateOf(MutableTransitionState(true))
+    }
+
+    /* 웹뷰 관련 */
     val isDarkMode = isDarkMode()
     val bridge = remember {
         CloseBridge(
@@ -111,15 +142,12 @@ fun LectureDetailPage() {
                 )
             }
         }
-    BackHandler(enabled = sheetState.isVisible) {
-        scope.launch { sheetState.hide() }
-    }
 
     CompositionLocalProvider(
         LocalReviewWebView provides reviewWebViewContainer,
     ) {
         ModalBottomSheetLayout(
-            sheetContent = sheetContent,
+            sheetContent = bottomSheetContent,
             sheetState = sheetState,
             sheetShape = RoundedCornerShape(topStartPercent = 5, topEndPercent = 5)
             // gesturesEnabled 가 없다! 그래서 드래그해서도 닫아진다..
@@ -128,6 +156,7 @@ fun LectureDetailPage() {
                 modifier = Modifier
                     .fillMaxSize()
                     .background(SNUTTColors.Gray100)
+                    .clicks { focusManager.clearFocus() }
             ) {
                 TopBar(
                     title = {
@@ -143,9 +172,16 @@ fun LectureDetailPage() {
                             modifier = Modifier
                                 .size(30.dp)
                                 .clicks {
-                                    if (editMode) editExitDialogState = true
-                                    else {
-                                        if (viewMode) vm.setViewMode(false)
+                                    if (editMode) {
+                                        editExitDialogState = true
+                                    } else if (vm.isAddMode()) {
+                                        vm.setAddMode(false)
+                                        vm.unsetEditMode()
+                                        navController.popBackStack()
+                                    } else if (viewMode) {
+                                        vm.setViewMode(false)
+                                        navController.popBackStack()
+                                    } else {
                                         navController.popBackStack()
                                     }
                                 },
@@ -163,12 +199,40 @@ fun LectureDetailPage() {
                                         if (editMode.not()) vm.setEditMode()
                                         else {
                                             scope.launch {
-                                                launchSuspendApi(apiOnProgress, apiOnError) {
-                                                    vm.updateLecture2()
-                                                    vm.initializeEditingLectureDetail(
-                                                        editingLectureDetail
-                                                    )
-                                                    vm.unsetEditMode()
+                                                lectureApiWithOverlapDialog(
+                                                    apiOnProgress,
+                                                    apiOnError,
+                                                    onLectureOverlap = { message ->
+                                                        lectureOverlapDialogMessage = message
+                                                        lectureOverlapDialogState = true
+                                                    }
+                                                ) {
+                                                    if (vm.isAddMode()) {
+                                                        vm.createLecture2()
+                                                        vm.unsetEditMode()
+                                                        vm.setAddMode(false)
+                                                        /* FIXME
+                                                             * 안드로이드는 여기서 그대로 detailPage 에 남는다.
+                                                             *
+                                                             * 하지만 @POST("/tables/{id}/lecture") api 는
+                                                             * tableDTO만 주기 때문에, 새로 추가된 커스텀 강의가 부여받은 id를
+                                                             * 알 수가 없다. (알려면 lectureList에서 find해야 되는데 nullable문제 및 compare 기준 문제 존재)
+                                                             * 그래서 바로 편집을 누르면 id를 모르는 강의에 대해 PUT을 하기 때문에 403이 난다.
+                                                             *
+                                                             * 그런데 기존 앱은 코드가 잘못 짜여져 있어서, 완료 후 바로 편집을 누르고
+                                                             * 완료를 다시 누르면 수정이 되는 게 아니라 또 추가가 된다. (계속 POST api를 쏜다)
+                                                             *
+                                                             * ios는 완료를 누르면 창이 닫히고 lecturesOfTable로 돌아가도록 돼 있다.
+                                                             * ios를 따라갈것인지, 서버 응답을 tableDto에서 lectureDto로 바꿔달라고 할 지 결정
+                                                             */
+                                                        scope.launch(Dispatchers.Main) { navController.popBackStack() }
+                                                    } else {
+                                                        vm.updateLecture2()
+                                                        vm.initializeEditingLectureDetail(
+                                                            editingLectureDetail
+                                                        )
+                                                        vm.unsetEditMode()
+                                                    }
                                                 }
                                             }
                                         }
@@ -229,7 +293,7 @@ fun LectureDetailPage() {
                                     ColorBox(
                                         editingLectureDetail.colorIndex,
                                         editingLectureDetail.color,
-                                        currentTable?.theme!!
+                                        tableColorTheme,
                                     )
                                     Spacer(modifier = Modifier.weight(1f))
                                     AnimatedVisibility(visible = editMode) {
@@ -246,106 +310,108 @@ fun LectureDetailPage() {
                     Margin(height = 10.dp)
                     Column(modifier = Modifier.background(SNUTTColors.White900)) {
                         Margin(height = 4.dp)
-                        LectureDetailItem(title = stringResource(R.string.lecture_detail_department)) {
-                            EditText(
-                                value = editingLectureDetail.department ?: "",
-                                onValueChange = {
-                                    vm.editEditingLectureDetail(editingLectureDetail.copy(department = it))
-                                },
-                                enabled = editMode,
-                                modifier = Modifier.fillMaxWidth(),
-                                underlineEnabled = false,
-                                textStyle = SNUTTTypography.body1.copy(fontSize = 15.sp),
-                                hint = stringResource(R.string.lecture_detail_hint_nothing),
-                            )
-                        }
-                        LectureDetailItem(title = stringResource(R.string.lecture_detail_academic_year)) {
-                            EditText(
-                                value = editingLectureDetail.academic_year ?: "",
-                                onValueChange = {
-                                    vm.editEditingLectureDetail(
-                                        editingLectureDetail.copy(
-                                            academic_year = it
+                        if (isCustom.not()) {
+                            LectureDetailItem(title = stringResource(R.string.lecture_detail_department)) {
+                                EditText(
+                                    value = editingLectureDetail.department ?: "",
+                                    onValueChange = {
+                                        vm.editEditingLectureDetail(editingLectureDetail.copy(department = it))
+                                    },
+                                    enabled = editMode,
+                                    modifier = Modifier.fillMaxWidth(),
+                                    underlineEnabled = false,
+                                    textStyle = SNUTTTypography.body1.copy(fontSize = 15.sp),
+                                    hint = stringResource(R.string.lecture_detail_hint_nothing),
+                                )
+                            }
+                            LectureDetailItem(title = stringResource(R.string.lecture_detail_academic_year)) {
+                                EditText(
+                                    value = editingLectureDetail.academic_year ?: "",
+                                    onValueChange = {
+                                        vm.editEditingLectureDetail(
+                                            editingLectureDetail.copy(
+                                                academic_year = it
+                                            )
                                         )
-                                    )
-                                },
-                                enabled = editMode,
-                                modifier = Modifier.fillMaxWidth(),
-                                underlineEnabled = false,
-                                textStyle = SNUTTTypography.body1.copy(fontSize = 15.sp),
-                                hint = stringResource(R.string.lecture_detail_hint_nothing),
-                            )
-                        }
-                        LectureDetailItem(title = stringResource(R.string.lecture_detail_credit)) {
-                            EditText(
-                                value = editingLectureDetail.credit.toString(),
-                                onValueChange = {
-                                    vm.editEditingLectureDetail(editingLectureDetail.copy(credit = it.stringToLong()))
-                                },
-                                enabled = editMode,
-                                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-                                modifier = Modifier.fillMaxWidth(),
-                                underlineEnabled = false,
-                                textStyle = SNUTTTypography.body1.copy(fontSize = 15.sp),
-                                hint = "0",
-                            )
-                        }
-                        LectureDetailItem(title = stringResource(R.string.lecture_detail_classification)) {
-                            EditText(
-                                value = editingLectureDetail.classification ?: "",
-                                onValueChange = {
-                                    vm.editEditingLectureDetail(
-                                        editingLectureDetail.copy(
-                                            classification = it
+                                    },
+                                    enabled = editMode,
+                                    modifier = Modifier.fillMaxWidth(),
+                                    underlineEnabled = false,
+                                    textStyle = SNUTTTypography.body1.copy(fontSize = 15.sp),
+                                    hint = stringResource(R.string.lecture_detail_hint_nothing),
+                                )
+                            }
+                            LectureDetailItem(title = stringResource(R.string.lecture_detail_credit)) {
+                                EditText(
+                                    value = editingLectureDetail.credit.toString(),
+                                    onValueChange = {
+                                        vm.editEditingLectureDetail(editingLectureDetail.copy(credit = it.stringToLong()))
+                                    },
+                                    enabled = editMode,
+                                    keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
+                                    modifier = Modifier.fillMaxWidth(),
+                                    underlineEnabled = false,
+                                    textStyle = SNUTTTypography.body1.copy(fontSize = 15.sp),
+                                    hint = "0",
+                                )
+                            }
+                            LectureDetailItem(title = stringResource(R.string.lecture_detail_classification)) {
+                                EditText(
+                                    value = editingLectureDetail.classification ?: "",
+                                    onValueChange = {
+                                        vm.editEditingLectureDetail(
+                                            editingLectureDetail.copy(
+                                                classification = it
+                                            )
                                         )
+                                    },
+                                    enabled = editMode,
+                                    modifier = Modifier.fillMaxWidth(),
+                                    underlineEnabled = false,
+                                    textStyle = SNUTTTypography.body1.copy(fontSize = 15.sp),
+                                    hint = stringResource(R.string.lecture_detail_hint_nothing),
+                                )
+                            }
+                            LectureDetailItem(title = stringResource(R.string.lecture_detail_category)) {
+                                EditText(
+                                    value = editingLectureDetail.category ?: "",
+                                    onValueChange = {
+                                        vm.editEditingLectureDetail(editingLectureDetail.copy(category = it))
+                                    },
+                                    enabled = editMode,
+                                    modifier = Modifier.fillMaxWidth(),
+                                    underlineEnabled = false,
+                                    textStyle = SNUTTTypography.body1.copy(fontSize = 15.sp),
+                                    hint = stringResource(R.string.lecture_detail_hint_nothing),
+                                )
+                            }
+                            LectureDetailItem(title = stringResource(R.string.lecture_detail_course_number)) {
+                                Text(
+                                    text = editingLectureDetail.course_number ?: "",
+                                    style = SNUTTTypography.body1.copy(
+                                        fontSize = 15.sp,
+                                        color = if (editMode) SNUTTColors.Gray200 else SNUTTColors.Black900
                                     )
-                                },
-                                enabled = editMode,
-                                modifier = Modifier.fillMaxWidth(),
-                                underlineEnabled = false,
-                                textStyle = SNUTTTypography.body1.copy(fontSize = 15.sp),
-                                hint = stringResource(R.string.lecture_detail_hint_nothing),
-                            )
-                        }
-                        LectureDetailItem(title = stringResource(R.string.lecture_detail_category)) {
-                            EditText(
-                                value = editingLectureDetail.category ?: "",
-                                onValueChange = {
-                                    vm.editEditingLectureDetail(editingLectureDetail.copy(category = it))
-                                },
-                                enabled = editMode,
-                                modifier = Modifier.fillMaxWidth(),
-                                underlineEnabled = false,
-                                textStyle = SNUTTTypography.body1.copy(fontSize = 15.sp),
-                                hint = stringResource(R.string.lecture_detail_hint_nothing),
-                            )
-                        }
-                        LectureDetailItem(title = stringResource(R.string.lecture_detail_course_number)) {
-                            Text(
-                                text = editingLectureDetail.course_number ?: "",
-                                style = SNUTTTypography.body1.copy(
-                                    fontSize = 15.sp,
-                                    color = if (editMode) SNUTTColors.Gray200 else SNUTTColors.Black900
                                 )
-                            )
-                        }
-                        LectureDetailItem(title = stringResource(R.string.lecture_detail_lecture_number)) {
-                            Text(
-                                text = editingLectureDetail.lecture_number ?: "",
-                                style = SNUTTTypography.body1.copy(
-                                    fontSize = 15.sp,
-                                    color = if (editMode) SNUTTColors.Gray200 else SNUTTColors.Black900
+                            }
+                            LectureDetailItem(title = stringResource(R.string.lecture_detail_lecture_number)) {
+                                Text(
+                                    text = editingLectureDetail.lecture_number ?: "",
+                                    style = SNUTTTypography.body1.copy(
+                                        fontSize = 15.sp,
+                                        color = if (editMode) SNUTTColors.Gray200 else SNUTTColors.Black900
+                                    )
                                 )
-                            )
-                        }
-                        LectureDetailItem(title = stringResource(R.string.lecture_detail_quota)) {
-                            Text(
-                                text = editingLectureDetail.quota.toString(),
-                                style = SNUTTTypography.body1.copy(
-                                    fontSize = 15.sp,
-                                    color = if (editMode) SNUTTColors.Gray200 else SNUTTColors.Black900
+                            }
+                            LectureDetailItem(title = stringResource(R.string.lecture_detail_quota)) {
+                                Text(
+                                    text = editingLectureDetail.quota.toString(),
+                                    style = SNUTTTypography.body1.copy(
+                                        fontSize = 15.sp,
+                                        color = if (editMode) SNUTTColors.Gray200 else SNUTTColors.Black900
+                                    )
                                 )
-                            )
+                            }
                         }
                         LectureDetailRemark(
                             title = stringResource(R.string.lecture_detail_remark),
@@ -377,19 +443,76 @@ fun LectureDetailPage() {
                             LectureDetailTimeAndLocation(
                                 timeText = SNUTTStringUtils.getClassTimeText(classTime),
                                 locationText = classTime.place,
+                                editTime = {
+                                    bottomSheetContentSetter.invoke {
+                                        DayTimePickerSheet(
+                                            classTime = classTime,
+                                            onDismiss = { scope.launch { sheetState.hide() } },
+                                            onConfirm = { editedClassTime ->
+                                                vm.editEditingLectureDetail(
+                                                    editingLectureDetail.copy(
+                                                        class_time_json = editingLectureDetail.class_time_json.toMutableList()
+                                                            .also {
+                                                                it[idx] = editedClassTime
+                                                            }
+                                                    )
+                                                )
+                                                scope.launch { sheetState.hide() }
+                                            }
+                                        )
+                                    }
+                                    scope.launch { sheetState.show() }
+                                },
                                 onLocationTextChange = { changedLocation ->
                                     vm.editEditingLectureDetail(
                                         editingLectureDetail.copy(
                                             class_time_json = editingLectureDetail.class_time_json.toMutableList()
                                                 .also {
-                                                    it[idx] =
-                                                        classTime.copy(place = changedLocation)
+                                                    it[idx] = classTime.copy(place = changedLocation)
                                                 }
                                         )
                                     )
                                 },
-                                editMode = editMode
+                                onClickDeleteIcon = {
+                                    editingClassTimeIndex = idx
+                                    deleteTimeDialogState = true
+                                },
+                                onLongClick = {
+                                    editingClassTimeIndex = idx
+                                    deleteTimeDialogState = true
+                                },
+                                editMode = editMode,
+                                isCustom = true,
+                                visible = if (idx == editingLectureDetail.class_time_json.lastIndex) classTimeAnimationState
+                                else MutableTransitionState(true)
                             )
+                        }
+                        AnimatedVisibility(
+                            visible = editMode
+                        ) {
+                            Box(
+                                modifier = Modifier
+                                    .fillMaxWidth()
+                                    .height(32.dp)
+                                    .clicks {
+                                        classTimeAnimationState =
+                                            MutableTransitionState(false).apply { targetState = true }
+                                        vm.editEditingLectureDetail(
+                                            editingLectureDetail.copy(
+                                                class_time_json = editingLectureDetail.class_time_json
+                                                    .toMutableList()
+                                                    .also { it.add(Defaults.defaultClassTimeDto) }
+                                            )
+                                        )
+                                    },
+                                contentAlignment = Alignment.Center
+                            ) {
+                                Text(
+                                    text = stringResource(R.string.lecture_detail_add_class_time),
+                                    textAlign = TextAlign.Center,
+                                    style = SNUTTTypography.body1.copy(color = SNUTTColors.Black600),
+                                )
+                            }
                         }
                         Margin(height = 7.dp)
                     }
@@ -398,40 +521,54 @@ fun LectureDetailPage() {
                     ) {
                         Column {
                             Margin(height = 30.dp)
-                            Column(modifier = Modifier.background(SNUTTColors.White900)) {
-                                LectureDetailButton(title = stringResource(R.string.lecture_detail_syllabus_button)) {
-                                    scope.launch {
-                                        launchSuspendApi(apiOnProgress, apiOnError) {
-                                            vm.getCourseBookUrl().let { url ->
-                                                val intent =
-                                                    Intent(Intent.ACTION_VIEW, Uri.parse(url))
-                                                context.startActivity(intent)
+                            if (isCustom) {
+                                Box(modifier = Modifier.background(Color.White)) {
+                                    LectureDetailButton(
+                                        title = stringResource(R.string.lecture_detail_delete_button),
+                                        textStyle = SNUTTTypography.body1.copy(
+                                            fontSize = 15.sp,
+                                            color = SNUTTColors.Red
+                                        )
+                                    ) {
+                                        deleteLectureDialogState = true
+                                    }
+                                }
+                            } else {
+                                Column(modifier = Modifier.background(SNUTTColors.White900)) {
+                                    LectureDetailButton(title = stringResource(R.string.lecture_detail_syllabus_button)) {
+                                        scope.launch {
+                                            launchSuspendApi(apiOnProgress, apiOnError) {
+                                                vm.getCourseBookUrl().let { url ->
+                                                    val intent =
+                                                        Intent(Intent.ACTION_VIEW, Uri.parse(url))
+                                                    context.startActivity(intent)
+                                                }
                                             }
                                         }
                                     }
-                                }
-                                LectureDetailButton(title = stringResource(R.string.lecture_detail_review_button)) {
-                                    scope.launch {
-                                        val job: CompletableJob = Job()
+                                    LectureDetailButton(title = stringResource(R.string.lecture_detail_review_button)) {
                                         scope.launch {
-                                            launchSuspendApi(apiOnProgress, apiOnError) {
-                                                reviewWebViewContainer.openPage(vm.getReviewContentsUrl() + "&on_back=close")
-                                                job.complete()
+                                            val job: CompletableJob = Job()
+                                            scope.launch {
+                                                launchSuspendApi(apiOnProgress, apiOnError) {
+                                                    reviewWebViewContainer.openPage(vm.getReviewContentsUrl() + "&on_back=close")
+                                                    job.complete()
+                                                }
                                             }
-                                        }
-                                        joinAll(job)
-                                        scope.launch {
-                                            sheetContentSetter.invoke {
-                                                ReviewWebView(0.95f)
+                                            joinAll(job)
+                                            scope.launch {
+                                                bottomSheetContentSetter.invoke {
+                                                    ReviewWebView(0.95f)
+                                                }
+                                                sheetState.show()
                                             }
-                                            sheetState.show()
                                         }
                                     }
                                 }
                             }
                         }
                     }
-                    if (vm.isViewMode().not()) {
+                    if (isCustom.not() && vm.isViewMode().not()) {
                         Margin(height = 10.dp)
                         Box(modifier = Modifier.background(Color.White)) {
                             LectureDetailButton(
@@ -505,6 +642,58 @@ fun LectureDetailPage() {
             title = stringResource(R.string.lecture_detail_reset_dialog_title)
         ) {
             Text(text = stringResource(R.string.lecture_detail_reset_dialog_message), style = SNUTTTypography.body2)
+        }
+    }
+
+    // 강의 시간대 삭제 다이얼로그
+    if (deleteTimeDialogState) {
+        CustomDialog(
+            onDismiss = { deleteTimeDialogState = false },
+            onConfirm = {
+                vm.editEditingLectureDetail(
+                    editingLectureDetail.copy(
+                        class_time_json = editingLectureDetail.class_time_json.toMutableList()
+                            .also {
+                                it.removeAt(editingClassTimeIndex)
+                            }
+                    )
+                )
+                deleteTimeDialogState = false
+            },
+            title = stringResource(R.string.lecture_detail_delete_class_time_message)
+        ) {}
+    }
+
+    // 강의 겹침 다이얼로그
+    if (lectureOverlapDialogState) {
+        CustomDialog(
+            onDismiss = { lectureOverlapDialogState = false },
+            onConfirm = {
+                scope.launch {
+                    if (vm.isAddMode()) {
+                        scope.launch {
+                            launchSuspendApi(apiOnProgress, apiOnError) {
+                                vm.createLecture2(is_forced = true)
+                                vm.unsetEditMode()
+                                vm.setAddMode(false)
+                                scope.launch(Dispatchers.Main) { navController.popBackStack() }
+                            }
+                        }
+                    } else {
+                        scope.launch {
+                            launchSuspendApi(apiOnProgress, apiOnError) {
+                                vm.updateLecture2(is_forced = true)
+                                vm.initializeEditingLectureDetail(editingLectureDetail)
+                                vm.unsetEditMode()
+                            }
+                        }
+                    }
+                    lectureOverlapDialogState = false
+                }
+            },
+            title = stringResource(id = R.string.lecture_overlap_error_message)
+        ) {
+            Text(text = lectureOverlapDialogMessage, style = SNUTTTypography.body2)
         }
     }
 }
@@ -714,6 +903,271 @@ fun ColorBox(
                 )
                 .size(20.dp)
         )
+    }
+}
+
+@Composable
+fun DayTimePickerSheet(
+    classTime: ClassTimeDto,
+    onDismiss: () -> Unit,
+    onConfirm: (ClassTimeDto) -> Unit,
+) {
+    val modalState = LocalModalState.current
+    val context = LocalContext.current
+    val dayList = remember { context.resources.getStringArray(R.array.week_days).map { it + "요일" } }
+    var dayIndex by remember { mutableStateOf(classTime.day) }
+
+    val amPmList = remember { listOf("오전", "오후") }
+    val hourList = remember { List(12) { if (it == 0) "12" else it.toString() } }
+    val minuteList = remember { List(12) { "%02d".format(it * 5) } }
+
+    var startTime: Time12 by remember { mutableStateOf(classTime.startTime12()) }
+    var endTime: Time12 by remember { mutableStateOf(classTime.endTime12()) }
+
+    var editingStartTime by remember { mutableStateOf(false) }
+    var editingEndTime by remember { mutableStateOf(false) }
+
+    /* 시작 시간이 끝나는 시간보다 같거나 더 나중일 때, 경계값 신경써서 조정하는 함수 */
+    val checkBoundary = {
+        if (startTime >= endTime) {
+            // 시작 시간을 끝나는 시간보다 나중으로 수정했으면, 끝나는 시간을 5분 뒤로 설정
+            if (editingStartTime) {
+                if (startTime.isLast()) {
+                    startTime = startTime.prev()
+                    endTime = startTime.next()
+                } else endTime = startTime.next()
+                // 끝나는 시간을 시작 시간보다 앞서게 수정했으면, 시작 시간을 5분 앞으로 설정
+            } else if (editingEndTime) {
+                if (endTime.isFirst()) {
+                    endTime = endTime.next()
+                    startTime = endTime.prev()
+                } else startTime = endTime.prev()
+            }
+        }
+    }
+
+    Column(
+        modifier = Modifier
+            .background(SNUTTColors.White900)
+            .padding(15.dp)
+    ) {
+        Row(modifier = Modifier.padding(5.dp)) {
+            Text(
+                text = stringResource(R.string.common_cancel),
+                style = SNUTTTypography.body1,
+                modifier = Modifier.clicks { onDismiss() },
+            )
+            Spacer(modifier = Modifier.weight(1f))
+            Text(
+                text = stringResource(R.string.common_ok),
+                style = SNUTTTypography.body1,
+                modifier = Modifier.clicks {
+                    onConfirm(
+                        classTime.copy(
+                            day = dayIndex,
+                            start_time = startTime.toString24(),
+                            end_time = endTime.toString24(),
+                        )
+                    )
+                }
+            )
+        }
+        Spacer(modifier = Modifier.height(30.dp))
+        Row(
+            modifier = Modifier
+                .padding(vertical = 7.dp)
+                .fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(text = "요일", style = SNUTTTypography.button)
+            Spacer(modifier = Modifier.weight(1f))
+            RoundBorderButton(
+                color = SNUTTColors.Gray400,
+                onClick = {
+                    var tempDayIndex by mutableStateOf(dayIndex)
+                    modalState
+                        .set(
+                            onDismiss = {
+                                dayIndex = tempDayIndex
+                                modalState.hide()
+                            },
+                            width = 150.dp,
+                        ) {
+                            Picker(
+                                list = dayList,
+                                initialCenterIndex = dayIndex,
+                                columnHeightDp = 45.dp,
+                                onValueChanged = { tempDayIndex = it }
+                            ) {
+                                Text(
+                                    text = dayList[it],
+                                    style = SNUTTTypography.button.copy(fontSize = 24.sp)
+                                )
+                            }
+                        }.show()
+                },
+            ) {
+                Text(text = dayList[dayIndex], style = SNUTTTypography.button)
+            }
+        }
+        Divider(color = SNUTTColors.Black250)
+        Row(
+            modifier = Modifier
+                .padding(vertical = 7.dp)
+                .fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(text = "시작", style = SNUTTTypography.button)
+            Spacer(modifier = Modifier.weight(1f))
+            RoundBorderButton(
+                color = SNUTTColors.Gray400,
+                onClick = {
+                    var tempStartTime by mutableStateOf(startTime.copy())
+                    editingStartTime = true
+                    modalState
+                        .set(
+                            onDismiss = {
+                                startTime = tempStartTime
+                                checkBoundary()
+                                editingStartTime = false
+                                modalState.hide()
+                            },
+                        ) {
+                            Row(verticalAlignment = Alignment.CenterVertically) {
+                                Box(modifier = Modifier.weight(1f)) {
+                                    Picker(
+                                        list = amPmList,
+                                        initialCenterIndex = startTime.amPm,
+                                        columnHeightDp = 45.dp,
+                                        onValueChanged = {
+                                            tempStartTime = tempStartTime.copy(amPm = it)
+                                        }
+                                    ) {
+                                        Text(
+                                            text = amPmList[it],
+                                            style = SNUTTTypography.button.copy(fontSize = 24.sp)
+                                        )
+                                    }
+                                }
+                                Box(modifier = Modifier.weight(1f)) {
+                                    CircularPicker(
+                                        list = hourList,
+                                        initialCenterIndex = startTime.hour,
+                                        columnHeightDp = 45.dp,
+                                        onValueChanged = {
+                                            tempStartTime = tempStartTime.copy(hour = it)
+                                        }
+                                    ) {
+                                        Text(
+                                            text = hourList[it],
+                                            style = SNUTTTypography.button.copy(fontSize = 24.sp)
+                                        )
+                                    }
+                                }
+                                Box(modifier = Modifier.weight(1f)) {
+                                    CircularPicker(
+                                        list = minuteList,
+                                        initialCenterIndex = startTime.minute / 5,
+                                        columnHeightDp = 45.dp,
+                                        onValueChanged = {
+                                            tempStartTime = tempStartTime.copy(minute = it * 5)
+                                        }
+                                    ) {
+                                        Text(
+                                            text = minuteList[it],
+                                            style = SNUTTTypography.button.copy(fontSize = 24.sp)
+                                        )
+                                    }
+                                }
+                            }
+                        }.show()
+                },
+            ) {
+                Text(
+                    text = startTime.toString(),
+                    style = SNUTTTypography.button
+                )
+            }
+        }
+        Divider(color = SNUTTColors.Black250)
+        Row(
+            modifier = Modifier
+                .padding(vertical = 7.dp)
+                .fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(text = "종료", style = SNUTTTypography.button)
+            Spacer(modifier = Modifier.weight(1f))
+            RoundBorderButton(
+                color = SNUTTColors.Gray400,
+                onClick = {
+                    var tempEndTime by mutableStateOf(endTime.copy())
+                    editingEndTime = true
+                    modalState
+                        .set(
+                            onDismiss = {
+                                endTime = tempEndTime
+                                checkBoundary()
+                                editingEndTime = false
+                                modalState.hide()
+                            },
+                        ) {
+                            Row(verticalAlignment = Alignment.CenterVertically) {
+                                Box(modifier = Modifier.weight(1f)) {
+                                    Picker(
+                                        list = amPmList,
+                                        initialCenterIndex = endTime.amPm,
+                                        columnHeightDp = 45.dp,
+                                        onValueChanged = {
+                                            tempEndTime = tempEndTime.copy(amPm = it)
+                                        }
+                                    ) {
+                                        Text(
+                                            text = amPmList[it],
+                                            style = SNUTTTypography.button.copy(fontSize = 24.sp)
+                                        )
+                                    }
+                                }
+                                Box(modifier = Modifier.weight(1f)) {
+                                    CircularPicker(
+                                        list = hourList,
+                                        initialCenterIndex = endTime.hour,
+                                        columnHeightDp = 45.dp,
+                                        onValueChanged = {
+                                            tempEndTime = tempEndTime.copy(hour = it)
+                                        }
+                                    ) {
+                                        Text(
+                                            text = hourList[it],
+                                            style = SNUTTTypography.button.copy(fontSize = 24.sp)
+                                        )
+                                    }
+                                }
+                                Box(modifier = Modifier.weight(1f)) {
+                                    CircularPicker(
+                                        list = minuteList,
+                                        initialCenterIndex = endTime.minute / 5,
+                                        columnHeightDp = 45.dp,
+                                        onValueChanged = {
+                                            tempEndTime = tempEndTime.copy(minute = it * 5)
+                                        }
+                                    ) {
+                                        Text(
+                                            text = minuteList[it],
+                                            style = SNUTTTypography.button.copy(fontSize = 24.sp)
+                                        )
+                                    }
+                                }
+                            }
+                        }.show()
+                },
+            ) {
+                Text(
+                    text = endTime.toString(),
+                    style = SNUTTTypography.button
+                )
+            }
+        }
     }
 }
