+++ app/src/main/java/com/wafflestudio/snutt2/views/logged_in/table_lectures/LecturesOfTablePage.kt
@@ -61,13 +61,12 @@ fun LecturesOfTablePage() {
                 lectureDetailViewModel.setAddMode(true)
                 lectureDetailViewModel.setEditMode()
                 lectureDetailViewModel.initializeEditingLectureDetail(Defaults.defaultLectureDto)
-                navController.navigate(NavigationDestination.LectureDetailCustom)
+                navController.navigate(NavigationDestination.LectureDetail)
             },
             onClickLecture = { lecture ->
                 // TODO: [자세히] 눌러서 들어간 화면은 조금 달라야 한다. (색깔, 편집 버튼, 삭제 버튼 없어야)
                 lectureDetailViewModel.initializeEditingLectureDetail(lecture)
-                if (lecture.isCustom) navController.navigate(NavigationDestination.LectureDetailCustom)
-                else navController.navigate(NavigationDestination.LectureDetail)
+                navController.navigate(NavigationDestination.LectureDetail)
             }
         )
     }
